Rename filenames: master→controller
