+++ src/main/core/controller.c
@@ -18,7 +18,7 @@
 #include <unistd.h>
 
 #include "main/core/logger/shadow_logger.h"
-#include "main/core/master.h"
+#include "main/core/controller.h"
 #include "main/core/manager.h"
 #include "main/core/support/configuration.h"
 #include "main/core/support/definitions.h"
