+++ src/main/core/main.c
@@ -19,7 +19,7 @@
 
 #include "external/elf-loader/dl.h"
 #include "main/core/logger/shadow_logger.h"
-#include "main/core/master.h"
+#include "main/core/controller.h"
 #include "main/core/support/configuration.h"
 #include "main/core/support/options.h"
 #include "main/host/affinity.h"
